Enforce strict warnings and add cppcheck/include-check tooling
Every first-party target -- the library, the ESP-IDF component, the unit
tests, and basic_player -- now compiles with the same warning set:
-Wconversion, -Wsign-conversion, -Wdouble-promotion, -Wformat=2,
-Wimplicit-fallthrough, and the Clang/GCC spelling of the dead-function
check, with -Werror behind the existing ENABLE_WERROR option so consumer
builds stay warning-only. The ESP component gets the block too: it is the
only build that compiles src/esp/ and the ESP_PLATFORM branches, so nothing
else can catch their warnings. -Wold-style-cast and -Wpedantic are left off
there because IDF's own headers trip them from inside our translation units
(pdMS_TO_TICKS, newlib's #include_next); both still apply on host.

The example's platformio.ini passes -DENABLE_WERROR=ON so CI's pinned ESP
build actually errors on those warnings.

Fallout fixed rather than silenced:
- audio_decoder: explicit casts on the FLAC header and samples_to_bytes()
  paths that were narrowing implicitly
- thread.h: esp_pthread_cfg_t::prio is unsigned; cast the int priority
- portaudio_sink: local no longer shadows free_space()

New tooling, mirroring the other micro audio libraries so the checks are
uniform across them:
- script/cppcheck.sh: whole-program pass whose unusedFunction check sees
  across translation units. Findings that only appear because cppcheck
  compiles the host branch with no codecs enabled carry inline suppressions
  with the reason; three genuinely dead test accessors are gone.
- script/check-includes.sh + check_includes.py + esp_stubs.py: IWYU via
  clang-include-cleaner, analyzing ESP branches against stub headers rather
  than a cross toolchain. src/host and src/esp are each checked in the one
  pass that matches how they compile. Fixed the includes it found missing
  (<utility>, <system_error>, <exception>, <cstdint>, esp_err.h,
  micro_decoder/types.h) and dropped an unused <cinttypes>.
- clang-tidy.sh honors a pre-set $CLANG_TIDY, validates the binary up
  front, and forces a non-zero exit on findings.

CI runs the two new checks as their own jobs and aggregates everything
into a single "CI success" status, which publish.yml now gates releases on
instead of matching job names by prefix. publish.yml also verifies the
manifest version against the release tag before publishing.

+++ src/audio_decoder.cpp
@@ -294,13 +294,14 @@ AudioDecoder::FileDecoderState AudioDecoder::decode_flac(const uint8_t* data, si
 
     if (result == micro_flac::FLAC_DECODER_SUCCESS) {
         if (samples_decoded > 0 && this->stream_info_.has_value()) {
-            this->output_buffer_.increase_length(
-                this->stream_info_.value().samples_to_bytes(samples_decoded));
+            this->output_buffer_.increase_length(this->stream_info_.value().samples_to_bytes(
+                static_cast<uint32_t>(samples_decoded)));
         }
     } else if (result == micro_flac::FLAC_DECODER_HEADER_READY) {
         const auto& info = this->flac_decoder_->get_stream_info();
         this->stream_info_ =
-            AudioStreamInfo(info.bits_per_sample(), info.num_channels(), info.sample_rate());
+            AudioStreamInfo(static_cast<uint8_t>(info.bits_per_sample()),
+                            static_cast<uint8_t>(info.num_channels()), info.sample_rate());
 
         this->free_buffer_required_ =
             static_cast<size_t>(this->flac_decoder_->get_output_buffer_size_samples()) *
@@ -494,8 +495,8 @@ AudioDecoder::FileDecoderState AudioDecoder::decode_wav(const uint8_t* data, siz
 
     if (result == micro_wav::WAV_DECODER_SUCCESS) {
         if (samples_decoded > 0 && this->stream_info_.has_value()) {
-            this->output_buffer_.increase_length(
-                this->stream_info_.value().samples_to_bytes(samples_decoded));
+            this->output_buffer_.increase_length(this->stream_info_.value().samples_to_bytes(
+                static_cast<uint32_t>(samples_decoded)));
         }
     } else if (result == micro_wav::WAV_DECODER_HEADER_READY) {
         this->stream_info_ =
